[AWIT-151] API generated
