+++ frontend/src/api/apis/forum-controller-api.ts
@@ -13,39 +13,25 @@
  */
 
 
-import globalAxios, {AxiosInstance, AxiosPromise, AxiosRequestConfig} from 'axios';
-import {Configuration} from '../configuration';
+import globalAxios, { AxiosPromise, AxiosInstance, AxiosRequestConfig } from 'axios';
+import { Configuration } from '../configuration';
 // Some imports not used depending on template conditions
 // @ts-ignore
-import {
-    assertParamExists,
-    createRequestFunction,
-    DUMMY_BASE_URL,
-    serializeDataIfNeeded,
-    setApiKeyToObject,
-    setBasicAuthToObject,
-    setBearerAuthToObject,
-    setOAuthToObject,
-    setSearchParams,
-    toPathString
-} from '../common';
+import { DUMMY_BASE_URL, assertParamExists, setApiKeyToObject, setBasicAuthToObject, setBearerAuthToObject, setOAuthToObject, setSearchParams, serializeDataIfNeeded, toPathString, createRequestFunction } from '../common';
 // @ts-ignore
-import {BASE_PATH, BaseAPI, COLLECTION_FORMATS, RequestArgs, RequiredError} from '../base';
+import { BASE_PATH, COLLECTION_FORMATS, RequestArgs, BaseAPI, RequiredError } from '../base';
 // @ts-ignore
+import { AddPostRequestBody } from '../models';
 // @ts-ignore
+import { AddThreadRequestBody } from '../models';
 // @ts-ignore
+import { ForumPost } from '../models';
 // @ts-ignore
+import { ForumPostUserIncluded } from '../models';
 // @ts-ignore
+import { ForumThread } from '../models';
 // @ts-ignore
-import {
-    AddPostRequestBody,
-    AddThreadRequestBody,
-    ForumPost,
-    ForumPostUserIncluded,
-    ForumThread,
-    ForumThreadSummaryResponseBody
-} from '../models';
-
+import { ForumThreadSummaryResponseBody } from '../models';
 /**
  * ForumControllerApi - axios parameter creator
  * @export
