+++ frontend/src/api/apis/plant-controller-api.ts
@@ -13,30 +13,21 @@
  */
 
 
-import globalAxios, {AxiosInstance, AxiosPromise, AxiosRequestConfig} from 'axios';
-import {Configuration} from '../configuration';
+import globalAxios, { AxiosPromise, AxiosInstance, AxiosRequestConfig } from 'axios';
+import { Configuration } from '../configuration';
 // Some imports not used depending on template conditions
 // @ts-ignore
-import {
-    assertParamExists,
-    createRequestFunction,
-    DUMMY_BASE_URL,
-    serializeDataIfNeeded,
-    setApiKeyToObject,
-    setBasicAuthToObject,
-    setBearerAuthToObject,
-    setOAuthToObject,
-    setSearchParams,
-    toPathString
-} from '../common';
+import { DUMMY_BASE_URL, assertParamExists, setApiKeyToObject, setBasicAuthToObject, setBearerAuthToObject, setOAuthToObject, setSearchParams, serializeDataIfNeeded, toPathString, createRequestFunction } from '../common';
 // @ts-ignore
-import {BASE_PATH, BaseAPI, COLLECTION_FORMATS, RequestArgs, RequiredError} from '../base';
+import { BASE_PATH, COLLECTION_FORMATS, RequestArgs, BaseAPI, RequiredError } from '../base';
 // @ts-ignore
+import { AddPlantRequestBody } from '../models';
 // @ts-ignore
+import { Plant } from '../models';
 // @ts-ignore
+import { PlantSummary } from '../models';
 // @ts-ignore
-import {AddPlantRequestBody, Plant, PlantsStats, PlantSummary} from '../models';
-
+import { PlantsStats } from '../models';
 /**
  * PlantControllerApi - axios parameter creator
  * @export
@@ -383,6 +374,44 @@ export const PlantControllerApiAxiosParamCreator = function (configuration?: Con
         },
         /**
          * 
+         * @summary Change sendReminders flag in plant
+         * @param {number} plantId 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        togglePlantReminders: async (plantId: number, options: AxiosRequestConfig = {}): Promise<RequestArgs> => {
+            // verify required parameter 'plantId' is not null or undefined
+            assertParamExists('togglePlantReminders', 'plantId', plantId)
+            const localVarPath = `/plants/{plantId}/toggle-reminders`
+                .replace(`{${"plantId"}}`, encodeURIComponent(String(plantId)));
+            // use dummy base URL string because the URL constructor only accepts absolute URLs.
+            const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL);
+            let baseOptions;
+            if (configuration) {
+                baseOptions = configuration.baseOptions;
+            }
+
+            const localVarRequestOptions = { method: 'PUT', ...baseOptions, ...options};
+            const localVarHeaderParameter = {} as any;
+            const localVarQueryParameter = {} as any;
+
+            // authentication JWT_AUTH required
+            // http bearer authentication required
+            await setBearerAuthToObject(localVarHeaderParameter, configuration)
+
+
+    
+            setSearchParams(localVarUrlObj, localVarQueryParameter);
+            let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {};
+            localVarRequestOptions.headers = {...localVarHeaderParameter, ...headersFromBaseOptions, ...options.headers};
+
+            return {
+                url: toPathString(localVarUrlObj),
+                options: localVarRequestOptions,
+            };
+        },
+        /**
+         * 
          * @summary Update plant
          * @param {number} plantId 
          * @param {AddPlantRequestBody} addPlantRequestBody 
@@ -534,6 +563,17 @@ export const PlantControllerApiFp = function(configuration?: Configuration) {
         },
         /**
          * 
+         * @summary Change sendReminders flag in plant
+         * @param {number} plantId 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        async togglePlantReminders(plantId: number, options?: AxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<void>> {
+            const localVarAxiosArgs = await localVarAxiosParamCreator.togglePlantReminders(plantId, options);
+            return createRequestFunction(localVarAxiosArgs, globalAxios, BASE_PATH, configuration);
+        },
+        /**
+         * 
          * @summary Update plant
          * @param {number} plantId 
          * @param {AddPlantRequestBody} addPlantRequestBody 
@@ -641,6 +681,16 @@ export const PlantControllerApiFactory = function (configuration?: Configuration
          */
         togglePlantFavourite(plantId: number, options?: any): AxiosPromise<void> {
             return localVarFp.togglePlantFavourite(plantId, options).then((request) => request(axios, basePath));
+        },
+        /**
+         * 
+         * @summary Change sendReminders flag in plant
+         * @param {number} plantId 
+         * @param {*} [options] Override http request option.
+         * @throws {RequiredError}
+         */
+        togglePlantReminders(plantId: number, options?: any): AxiosPromise<void> {
+            return localVarFp.togglePlantReminders(plantId, options).then((request) => request(axios, basePath));
         },
         /**
          * 
@@ -767,6 +817,18 @@ export class PlantControllerApi extends BaseAPI {
      */
     public togglePlantFavourite(plantId: number, options?: AxiosRequestConfig) {
         return PlantControllerApiFp(this.configuration).togglePlantFavourite(plantId, options).then((request) => request(this.axios, this.basePath));
+    }
+
+    /**
+     * 
+     * @summary Change sendReminders flag in plant
+     * @param {number} plantId 
+     * @param {*} [options] Override http request option.
+     * @throws {RequiredError}
+     * @memberof PlantControllerApi
+     */
+    public togglePlantReminders(plantId: number, options?: AxiosRequestConfig) {
+        return PlantControllerApiFp(this.configuration).togglePlantReminders(plantId, options).then((request) => request(this.axios, this.basePath));
     }
 
     /**
