+++ frontend/src/api/models/auth-response.ts
@@ -13,7 +13,7 @@
  */
 
 
-import {AuthData} from './auth-data';
+import { AuthData } from './auth-data';
 
 /**
  *
